edit home tab
