+++ app/src/main/java/com/lifeistech/android/weatherforecast/MainActivity.java
@@ -29,8 +29,6 @@ import rx.Observer;
 import rx.schedulers.Schedulers;
 
 public class MainActivity extends Activity {
-	private static final String END_POINT = "http://weather.livedoor.com";
-	private static final String TAG = MainActivity.class.getSimpleName();
 
 	private Tab tab;
 
@@ -82,49 +80,4 @@ public class MainActivity extends Activity {
 		getMenuInflater().inflate(R.menu.activity_main, menu);
 		return true;
 	}
-
-	public void retrofitConnect() {
-		Gson gson = new GsonBuilder()
-				.setFieldNamingPolicy(FieldNamingPolicy.LOWER_CASE_WITH_UNDERSCORES)
-				.registerTypeAdapter(Date.class, new DateTypeAdapter())
-				.create();
-
-
-		// RestAdapterを作成する
-		RestAdapter adapter = new RestAdapter.Builder()
-				.setEndpoint(END_POINT)
-				.setConverter(new GsonConverter(gson))
-				.setLogLevel(RestAdapter.LogLevel.FULL)
-				.setLog(new AndroidLog("=NETWORK="))
-				.build();
-
-		// 天気予報情報を取得する
-		//http://weather.livedoor.com/area/forecast/200010
-		WeatherApi api =  adapter.create(WeatherApi.class);
-
-		Observer observer = new Observer<WeatherEntity>() {
-			@Override
-			public void onCompleted() {
-				Log.d(TAG, "onCompleted()");
-				//必要な情報を取り出して画面に表示してください。
-			}
-
-			@Override
-			public void onError(Throwable e) {
-				Log.e(TAG, "Error : " + e.toString());
-			}
-
-			@Override
-			public void onNext(WeatherEntity weather) {
-				Log.d(TAG, "onNext()");
-				((TextView)findViewById(R.id.text)).setText(weather.getPinpointLocations().get(0).getName());
-			}
-		};
-
-		api.getWeather("200010")
-				.subscribeOn(Schedulers.newThread())
-				.observeOn(AndroidSchedulers.mainThread())
-				.subscribe(observer);
-	}
-
 }
